pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PeakDeepMaster"
version = "0.1.0"
authors = [
  { name="Diego Baron" },
]
description = "Coupling estimation with signed weights using PyTorch Lightning."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "h5py",
  "matplotlib",
  "scikit-learn",
  "hydra-core",
  "rich",
  "torch",
  "torchvision",
  "torchaudio",
  "lightning>=2.5.0,<2.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
